Look where you go control scheme
Look on the left or right edge of the screen to turn the camera, look bottom left to move down and bottom right to move up; fixate on a point for at least 2 seconds to have the drone go there. There is momentum with the fixating action (if you fixate multiple times without turning the camera or moving up/down you will go further).